move results dir to a new name to avoid colliding with opics_common.results at runtime
