+++ core/test_register.py
@@ -6,7 +6,7 @@ from opics_common.launch.opics_run_state        import NOT_ATTEMPTED, IN_PROGRES
 from core.constants                 import JOB_REQUEST, JOB_REQUEST_SMOKE, JOB_ASSIGN, NO_MORE_SCENES_TO_RUN, SESSION_KILLED, SMOKE_TEST
 from core.constants                 import TEST_HISTORY_FIRST_LINE_PREFIX
 from core.optics_session            import OpticsSession
-from results.scene_state_history    import SceneStateHistory
+from optics_results.scene_state_history    import SceneStateHistory
 from core.utils                     import optics_info, optics_error, optics_debug, optics_fatal
